pyproject.toml
[project]
name = "ig-ctf-solver"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.5.0",
    "factordb-pycli>=1.3.0",
    "flask>=3.1.2",
    "gmpy2>=2.3.0",
    "inspect-ai>=0.3.179",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.9",
    "langgraph>=1.0.8",
    "numpy>=2.4.2",
    "pandas>=3.0.0",
    "primefac>=2.0.12",
    "pwntools>=4.15.0",
    "pycryptodomex>=3.23.0",
    "pydantic-settings>=2.13.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "ropgadget>=7.7",
    "sentence-transformers>=5.2.2",
    "sympy>=1.14.0",
    "tqdm>=4.67.3",
    "z3-solver>=4.15.8.0",
]
